feat: notifications logic hook
